Create Words Dictionary using Linkedlist
Step1: Download the txt file from mentioned google drive link and save it in the same folder where you will save the code of this repo. 
Step2: Create .py file and paste the code. and then run it

Note:
This is only code which will fetch data (words and their meanings) from txt file and saves each word with its meaning in a node of linkedlist, and will create number of nodes depending upon words in txt file